restore previously reverted code

+++ arches/app/views/concept.py
@@ -44,16 +44,17 @@ def rdm(request, conceptid):
 @csrf_exempt
 def concept(request, conceptid):
     ret = {'success': False}
+    f = request.GET.get('f', 'json')
+    lang = request.GET.get('lang', 'en-us')
+    pretty = request.GET.get('pretty', False)
 
     if request.method == 'GET':
         include_subconcepts = request.GET.get('include_subconcepts', 'true') == 'true'
         include_parentconcepts = request.GET.get('include_parentconcepts', 'true') == 'true'
+        include_relatedconcepts = request.GET.get('include_relatedconcepts', 'true') == 'true'
         emulate_elastic_search = request.GET.get('emulate_elastic_search', 'false') == 'true'
         fromdb = request.GET.get('fromdb', 'false') == 'true'
-        f = request.GET.get('f', 'json')
-        lang = request.GET.get('lang', 'en-us')
         depth_limit = request.GET.get('depth_limit', None)
-        pretty = request.GET.get('pretty', False)
 
         if f == 'html':
             fromdb = True
@@ -66,7 +67,8 @@ def concept(request, conceptid):
 
         if fromdb:
             concept_graph = Concept().get(id=conceptid, include_subconcepts=include_subconcepts, 
-                include_parentconcepts=include_parentconcepts, depth_limit=depth_limit, up_depth_limit=None)
+                include_parentconcepts=include_parentconcepts, include_relatedconcepts=include_relatedconcepts,
+                depth_limit=depth_limit, up_depth_limit=None)
             
             nodes = [{'concept_id': concept_graph.id, 'name': concept_graph.get_preflabel(lang=lang).value,'type': 'Current'}]
             links = []
@@ -174,7 +176,7 @@ def concept(request, conceptid):
                 ret = concept.graph(include_subconcepts=False, include_parentconcepts=False, include=['label'])
                 return JSONResponse(ret, indent=(4 if request.GET.get('pretty', False) else None))
 
-            elif data['action'] == 'manage_related_concept':
+            elif data['action'] == 'manage-related-concept':
                 relation = None
                 if 'related_concept' in data:
                     relation = archesmodels.ConceptRelations()
@@ -196,19 +198,22 @@ def concept(request, conceptid):
 
 
     if request.method == 'DELETE':
-            ret = {
-                'deleted': [],
-                'success': False
-            }
-            concept = archesmodels.Concepts.objects.get(conceptid = conceptid)
-            concepts = concept.graph(exclude=['note','label',None], include_parentconcepts=False).flatten()
+        json = request.body
+        if json != None:
+            data = JSONDeserializer().deserialize(json)
+            if 'action' in data:
+                action = data['action']
 
-            with transaction.atomic():
-                for concept in concepts:
-                    ret['deleted'].append(concept.get_preflabel())
+                if data['action'] == 'delete-relationship':
+                    relatedconceptid = data['relatedconceptid']
+                    concept = Concept({'id':conceptid})
+                    concept.addrelatedconcept({'id': relatedconceptid})
+                    concept.delete_related_concept()
+                
+                elif data['action'] == 'delete-concept':
+                    concept = Concept({'id':conceptid})
                     concept.delete()
-
-            ret['success'] = True
+                    ret['success'] = True
 
     return JSONResponse(ret, indent=(4 if pretty else None))
 
@@ -261,55 +266,5 @@ def search(request):
 
 def concept_tree(request):
     conceptid = request.GET.get('node', None)
-    top_concept = '00000000-0000-0000-0000-000000000003'
-    class concept(object):
-        def __init__(self, *args, **kwargs):
-            self.label = ''
-            self.labelid = ''
-            self.id = ''
-            self.load_on_demand = False
-            self.children = []    
-
-    def _findNarrowerConcept(conceptid, depth_limit=None, level=0):
-        labels = archesmodels.Values.objects.filter(conceptid = conceptid)
-        ret = concept()          
-        for label in labels:
-            if label.valuetype_id == 'prefLabel':
-                ret.label = label.value
-                ret.id = label.conceptid_id
-                ret.labelid = label.valueid
-
-        conceptrealations = archesmodels.ConceptRelations.objects.filter(conceptidfrom = conceptid)
-        if depth_limit != None and len(conceptrealations) > 0 and level >= depth_limit:
-            ret.load_on_demand = True
-        else:
-            if depth_limit != None:
-                level = level + 1                
-            for relation in conceptrealations:
-                ret.children.append(_findNarrowerConcept(relation.conceptidto_id, depth_limit=depth_limit, level=level))   
-
-        return ret 
-
-    def _findBroaderConcept(conceptid, child_concept, depth_limit=None, level=0):
-        conceptrealations = archesmodels.ConceptRelations.objects.filter(conceptidto = conceptid)
-        if len(conceptrealations) > 0 and conceptid != top_concept:
-            labels = archesmodels.Values.objects.filter(conceptid = conceptrealations[0].conceptidfrom_id)
-            ret = concept()          
-            for label in labels:
-                if label.valuetype_id == 'prefLabel':
-                    ret.label = label.value
-                    ret.id = label.conceptid_id
-                    ret.labelid = label.valueid
-
-            ret.children.append(child_concept)
-            return _findBroaderConcept(conceptrealations[0].conceptidfrom_id, ret, depth_limit=depth_limit, level=level)
-        else:
-            return child_concept
-    
-    if conceptid == None or conceptid == '':
-        concepts = [_findNarrowerConcept(top_concept, depth_limit=1)]
-    else:
-        concepts = _findNarrowerConcept(conceptid, depth_limit=1)
-        concepts = [_findBroaderConcept(conceptid, concepts, depth_limit=1)]
-
+    concepts = Concept({'id': conceptid}).concept_tree(top_concept = '00000000-0000-0000-0000-000000000003')
     return JSONResponse(concepts, indent=4)
